Add fitness function

+++ src/main/board.h
@@ -67,7 +67,7 @@ class Board{
          * @brief This will be the 'heuristic function'
          * Check average violations - current violations.
          * return value from that
-         * @return long 
+         * @return double 
          */
-        long fitnessFunction(size_t);
+        double fitnessFunction(double);
 };
